+++ src/loader/updaters/CustomDownloader.h
@@ -13,7 +13,7 @@ namespace loader::updaters {
 
     protected:
         virtual void DownloadUpdate() override;
-        static void GW2ADDON_CALL WriteBuffer(void* downloader, size_t progress, size_t totalSize);
+        static void GW2ADDON_CALL WriteBuffer(void* downloader, unsigned int progress, unsigned int totalSize);
 
     private:
         GW2AddonDownloadUpdate_t* downloadFunc;
